upnp: fix for compiler warnings and incorrect API.

+++ upnp/inc/upnp.h
@@ -1098,7 +1098,7 @@ EXPORT_SPEC int UpnpSetContentLength(
 	 * length needs to be set. */
 	UpnpClient_Handle Hnd,
 	/*! [in] Permissible content length */
-	int contentLength);
+	size_t contentLength);
 
 
 /*!
